feat: code coloring & find source file
